fix: :rotating_light: fix broken code
