--- package.json
@@ -1,16 +1,0 @@
-{
-  "name": "th-projekt-3",
-  "version": "0.1.0",
-  "dependencies": {
-    "openseadragon": "^2.4.2",
-    "vue": "^3.2.22",
-    "vue-router": "^4.0.0-0",
-    "vuedraggable": "^4.0.3"
-  },
-  "devDependencies": {
-    "@types/openseadragon": "^2.4.5",
-    "sass": "^1.26.5",
-    "sass-loader": "^8.0.2",
-    "typescript": "~4.1.5"
-  }
-}
